fix: updating branch and reworking redirections;
