+++ outfile
@@ -1,11 +1,3 @@
-here? 4
-here? 5
-here? 6
-here? 7
-here? 4
-here? 5
-here? 6
-here? 7
 test
